+++ CHANGELOG.md
@@ -14,9 +14,9 @@ Global Functions
 
 Macros
 
-- Macros are added to `VCore`
 - `CaseDetection` macro is added, that adds computed properties to `enum` to represent each `case`
-- `MemberwiseCodable` and `MWCKey` macros are added, that add `CodingKeys` to a declaration to memberwise code each property.
+- `MemberwiseCodable` and `MWCKey` macros are added, that add `CodingKeys` to a declaration to memberwise code each property
+- `OptionSetRepresentation` macro is added, that generates an `OptionSet` from a `struct` that contains a nested `Options` `enum`
 - `URL(_:)` macro is added, that creates a non-optional `URL` from a `String` literal
 
 ### [5.2.1(88)](https://github.com/VakhoKontridze/VCore/releases/tag/5.2.1) — *2023 12 08*